distconf - Fallback to last known good value on invalid config
The old code used to continue looking for valid config values
in lower backing stores when an update at a higher level
contained an invalid value. This could cause a simple typo to
revert a distconf variable to another state, which could
be bad(tm). Now on an invalid value the current value of the
config is kept.

Also refactored distconf unit test so that adding a new type
will be easier to test.